asfasfd

+++ MySql/API/Items/ItemsAPI.js
@@ -1,7 +1,8 @@
 const menuItemsMethods = require('../MenuItems/Methods');
-const announcemtMethods = require('../Announcements/Methods');
+const announcementMethods = require('../Announcements/Methods');
 const pageMethods = require('../PageItems/Methods');
 const mediaMethods = require('../MediaItems/Methods');
+const categoriesMethods = require('../Categories/Methods');
 
 async function deleteItem(req, res, next) {
   var serverResponse = null;
@@ -10,12 +11,24 @@ async function deleteItem(req, res, next) {
     rows = await menuItemsMethods.deleteItem(req, res, next);
   else if (req.body.kind == 2)
     rows = serverResponse = await pageMethods.deleteItem(req, res, next);
-  else if (req.body.kind == 3)
-    rows = serverResponse = await mediaMethods.deleteItem(req, res, next);
-  else if (req.body.kind == 4)
-    rows = serverResponse = await announcemtMethods.deleteItem(req, res, next);
+  else if (req.body.kind == 3) {
+    var ret = await mediaMethods.deleteMediaFromDirectory(req, res, next);
+    if (ret) {
+      var serverResponse = await mediaMethods.deleteItem(req, res, next);
+      if (serverResponse.affectedRows === 1)
+        res.status(200).json(req.body);
+      else
+        res.status(205).json(undefined);
+    } else
+      res.status(450).json('Internal Server Error');
 
-  if (rows && rows.affectedRows > 0)  
+
+  } else if (req.body.kind == 4)
+    rows = serverResponse = await announcementMethods.deleteItem(req, res, next);
+  else if (req.body.kind == 5)
+    rows = serverResponse = await categoriesMethods.deleteItem(req, res, next);
+
+  if (rows && rows.affectedRows > 0)
     res.status(200).json(req.body);
   else
     res.status(709).json(serverResponse);
